Completed tutorilal 5
Learned about migrations with the manage.py file and adding to the database via the shell.

How to query the data from db to be used for context when rendering pages. Would be interested to find out how
to do this with plots. I.e integrate plotly with this.

Also learned about accessing and editing posts by registering them in the admin page.